requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.14
django-crispy-forms==1.11.2
gunicorn==20.1.0
lxml==4.9.1
psycopg2-binary==2.8.6
python-docx==0.8.10
pytz==2021.1
sqlparse==0.4.2
